# packages/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Montessori ERP & LMS API
  description: |
    Multi-tenant, role-based ERP & Learning Management System for Montessori schools.
    
    ## Authentication
    All routes (except `/api/v1/auth/*`) require a JWT Bearer token in the `Authorization` header.
    
    ## Roles
    - **admin**: Full access to all modules
    - **teacher**: Students, attendance, observations, curriculum, messages
    - **parent**: Read-only access to own children, fees, messages
    - **student**: Read-only curriculum and announcements
  version: 1.0.0
  contact:
    name: Montessori ERP Support
servers:
  - url: http://localhost:4001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
...
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new school (tenant + admin user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schoolName, subdomain, adminEmail, adminPassword]
              properties:
                schoolName:
                  type: string
                subdomain:
                  type: string
                adminEmail:
                  type: string
                  format: email
                adminPassword:
                  type: string
                  minLength: 8
                adminFirstName:
                  type: string
                adminLastName:
                  type: string
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tenant:
                    type: object
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '409':
          description: School already exists
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New access token
        '401':
          description: Invalid refresh token
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (invalidate refresh token)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful

  # ==================== STUDENTS ====================
...
  /api/v1/classes:
    get:
      tags: [Classes]
      summary: List classrooms
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated classrooms
    post:
      tags: [Classes]
      summary: Create a classroom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, academicYear]
              properties:
                name:
                  type: string
                academicYear:
                  type: string
                teacherIds:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Classroom created
  /api/v1/classes/{id}:
    get:
      tags: [Classes]
      summary: Get classroom by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Classroom details
    patch:
      tags: [Classes]
      summary: Update classroom
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Classroom updated
    delete:
      tags: [Classes]
      summary: Delete classroom
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Classroom deleted

  # ==================== ATTENDANCE ====================
  /api/v1/attendance:
    get:
      tags: [Attendance]
      summary: List attendance records
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: classId
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Attendance records
    post:
      tags: [Attendance]
      summary: Record attendance (supports offline batch upsert)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, date, status]
              properties:
                studentId:
                  type: string
                date:
                  type: string
                  format: date
                status:
                  type: string
                  enum: [PRESENT, ABSENT, LATE, EXCUSED]
                reasonCode:
                  type: string
      responses:
        '201':
          description: Attendance recorded
        '200':
          description: Batch upsert complete
  /api/v1/attendance/bulk:
    post:
      tags: [Attendance]
      summary: Batch upsert attendance (offline sync)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                records:
                  type: array
                  items:
                    type: object
                    properties:
                      studentId:
                        type: string
                      date:
                        type: string
                      status:
                        type: string
      responses:
        '200':
          description: Batch complete
  /api/v1/attendance/{id}:
    patch:
      tags: [Attendance]
      summary: Update attendance record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attendance updated

  # ==================== OBSERVATIONS ====================
...